Add framed tranport option to server
Many async thrift servers used the "framed" protocol, which is incompatable with the unframed version and has to be set via configuration.  People interested in aiothrift may be migrating from eg. the Python 2 Twisted bindings which support only the framed protocol.  Therefore adding this option makes it a drop-in replacement for those servers.